Added basic statistics

+++ Tools.h
@@ -25,8 +25,8 @@ inline int positive_modulo(int i, int n) {
     return (i % n + n) % n;
 }
 
-inline void displayText(const float &x, const float &y, const int &r,
-        const int &g, const int &b, const std::string &string) {
+inline void displayText(const float &x, const float &y, const float &r,
+        const float &g, const float &b, const std::string &string) {
     glColor3f(r, g, b);
     glRasterPos2f(x, y);
     for (unsigned int i = 0; i < string.length(); i++) {
